analysis-service/requirements.txt
flask==2.3.2
openai==1.3.0
transformers==4.31.0
torch==2.0.1
numpy==1.25.0
pandas==2.0.3
spacy==3.6.0
scikit-learn==1.3.0
python-docx==0.8.11
pdfminer.six==20221105
pymongo==4.4.1
python-dotenv==1.0.0
gunicorn==21.2.0
pydantic==2.0.3
pytest==7.4.0
requests==2.31.0
